#68 メインクラスの機能を最小化
