+++ src/technote.php
@@ -47,7 +47,6 @@ define( 'TECHNOTE_IS_MOCK', false );
  * @property string $plugin_name
  * @property string $slug_name
  * @property string $plugin_file
- * @property array $plugin_data
  * @property \Technote\Classes\Models\Lib\Define $define
  * @property \Technote\Classes\Models\Lib\Config $config
  * @property \Technote\Classes\Models\Lib\Setting $setting
@@ -69,17 +68,29 @@ define( 'TECHNOTE_IS_MOCK', false );
  * @property \Technote\Classes\Models\Lib\Social $social
  * @property \Technote\Classes\Models\Lib\Custom_Post $custom_post
  * @property \Technote\Classes\Models\Lib\Mail $mail
+ * @method void main_init( bool $uninstall = false )
+ * @method bool has_initialized()
+ * @method string get_plugin_version()
+ * @method string|false get_text_domain()
+ * @method string translate( string $value )
+ * @method mixed get_config( string $name, string $key, mixed $default = null )
+ * @method mixed get_option( string $key, mixed $default = '' )
+ * @method mixed get_session( string $key, mixed $default = '' )
+ * @method mixed set_session( string $key, mixed $value, int | null $duration = null )
+ * @method bool user_can( null | string | false $capability = null )
+ * @method void log( string $message, mixed $context = null, string $level = '' )
+ * @method void add_message( string $message, string $group = '', bool $error = false, bool $escape = true )
+ * @method string get_page_slug( string $file )
+ * @method mixed get_shared_object( string $key, string | null $target = null )
+ * @method void set_shared_object( string $key, mixed $object, string | null $target = null )
+ * @method bool isset_shared_object( string $key, string | null $target = null )
+ * @method void delete_shared_object( string $key, string | null $target = null )
+ * @method array|string get_plugin_data( string | null $key = null )
  */
 class Technote {
 
 	/** @var array|\Technote[] */
 	private static $instances = [];
-
-	/** @var bool $lib_language_loaded */
-	private static $lib_language_loaded = false;
-
-	/** @var array $shared_object */
-	private static $shared_object = [];
 
 	/**
 	 * @since 2.0.0
@@ -92,15 +103,6 @@ class Technote {
 	 * @var string $latest_library_directory
 	 */
 	private static $latest_library_directory = null;
-
-	/**
-	 * @since 2.0.0
-	 * @var bool $plugins_loaded
-	 */
-	private $plugins_loaded = false;
-
-	/** @var bool $initialized */
-	private $initialized = false;
 
 	/**
 	 * @since 2.0.0
@@ -123,40 +125,37 @@ class Technote {
 	/** @var string $plugin_file */
 	public $plugin_file;
 
-	/** @var array $plugin_data */
-	public $plugin_data;
+	/** @var string $slug_name */
+	public $slug_name;
 
 	/**
-	 * @since 2.8.0 Added: social, custom_post
-	 * @since 2.9.0 Added: mail
-	 * @var array $properties
+	 * @since 2.0.0
+	 * @var bool $plugins_loaded
 	 */
-	private $properties = [
-		'define'      => '\Technote\Classes\Models\Lib\Define',
-		'config'      => '\Technote\Classes\Models\Lib\Config',
-		'setting'     => '\Technote\Classes\Models\Lib\Setting',
-		'option'      => '\Technote\Classes\Models\Lib\Option',
-		'device'      => '\Technote\Classes\Models\Lib\Device',
-		'minify'      => '\Technote\Classes\Models\Lib\Minify',
-		'filter'      => '\Technote\Classes\Models\Lib\Filter',
-		'user'        => '\Technote\Classes\Models\Lib\User',
-		'post'        => '\Technote\Classes\Models\Lib\Post',
-		'loader'      => '\Technote\Classes\Models\Lib\Loader',
-		'log'         => '\Technote\Classes\Models\Lib\Log',
-		'input'       => '\Technote\Classes\Models\Lib\Input',
-		'db'          => '\Technote\Classes\Models\Lib\Db',
-		'uninstall'   => '\Technote\Classes\Models\Lib\Uninstall',
-		'session'     => '\Technote\Classes\Models\Lib\Session',
-		'utility'     => '\Technote\Classes\Models\Lib\Utility',
-		'test'        => '\Technote\Classes\Models\Lib\Test',
-		'upgrade'     => '\Technote\Classes\Models\Lib\Upgrade',
-		'social'      => '\Technote\Classes\Models\Lib\Social',
-		'custom_post' => '\Technote\Classes\Models\Lib\Custom_Post',
-		'mail'        => '\Technote\Classes\Models\Lib\Mail',
-	];
+	private $plugins_loaded = false;
 
-	/** @var array $property_instances */
-	private $property_instances = [];
+	/**
+	 * @since 2.10.0
+	 * @var \Technote\Classes\Models\Lib\Main $main
+	 */
+	private $main;
+
+	/**
+	 * Technote constructor.
+	 *
+	 * @param string $plugin_name
+	 * @param string $plugin_file
+	 * @param string|null $slug_name
+	 */
+	private function __construct( $plugin_name, $plugin_file, $slug_name ) {
+		$this->original_plugin_name = $plugin_name;
+		$this->plugin_file          = $plugin_file;
+		$this->plugin_name          = strtolower( $this->original_plugin_name );
+		$this->slug_name            = ! empty( $slug_name ) ? strtolower( $slug_name ) : $this->plugin_name;
+
+		$this->setup_library_version();
+		$this->setup_actions();
+	}
 
 	/**
 	 * @param string $name
@@ -165,16 +164,7 @@ class Technote {
 	 * @throws \OutOfRangeException
 	 */
 	public function __get( $name ) {
-		if ( isset( $this->properties[ $name ] ) ) {
-			if ( ! isset( $this->property_instances[ $name ] ) ) {
-				/** @var \Technote\Interfaces\Singleton $class */
-				$class                             = $this->properties[ $name ];
-				$this->property_instances[ $name ] = $class::get_instance( $this );
-			}
-
-			return $this->property_instances[ $name ];
-		}
-		throw new \OutOfRangeException( $name . ' is undefined.' );
+		return $this->get_main()->__get( $name );
 	}
 
 	/**
@@ -183,7 +173,55 @@ class Technote {
 	 * @return bool
 	 */
 	public function __isset( $name ) {
-		return array_key_exists( $name, $this->properties );
+		return $this->get_main()->__isset( $name );
+	}
+
+	/**
+	 * @param $name
+	 * @param $arguments
+	 *
+	 * @return mixed
+	 */
+	public function __call( $name, $arguments ) {
+		return $this->get_main()->$name( ...$arguments );
+	}
+
+	/**
+	 * @param $name
+	 * @param $arguments
+	 */
+	public static function __callStatic( $name, $arguments ) {
+		if ( preg_match( '#register_uninstall_(.+)\z#', $name, $matches ) ) {
+			$plugin_base_name = $matches[1];
+			self::uninstall( $plugin_base_name );
+		}
+	}
+
+	/**
+	 * @since 2.10.0
+	 * @return \Technote\Classes\Models\Lib\Main|\Technote\Traits\Singleton
+	 */
+	private function get_main() {
+		if ( ! isset( $this->main ) ) {
+			$required = [
+				'Interfaces\Readonly',
+				'Interfaces\Singleton',
+				'Traits\Readonly',
+				'Traits\Singleton',
+				'Classes\Models\Lib\Main',
+			];
+			$dir      = self::$latest_library_directory . DS . 'src';
+			foreach ( $required as $item ) {
+				$path = $dir . DS . str_replace( '\\', DS, strtolower( $item ) ) . '.php';
+				if ( is_readable( $path ) ) {
+					/** @noinspection PhpIncludeInspection */
+					require_once $path;
+				}
+			}
+			$this->main = \Technote\Classes\Models\Lib\Main::get_instance( $this );
+		}
+
+		return $this->main;
 	}
 
 	/**
@@ -207,23 +245,6 @@ class Technote {
 		}
 
 		return self::$instances[ $plugin_name ];
-	}
-
-	/**
-	 * Technote constructor.
-	 *
-	 * @param string $plugin_name
-	 * @param string $plugin_file
-	 * @param string|null $slug_name
-	 */
-	private function __construct( $plugin_name, $plugin_file, $slug_name ) {
-		$this->original_plugin_name = $plugin_name;
-		$this->plugin_file          = $plugin_file;
-		$this->plugin_name          = strtolower( $this->original_plugin_name );
-		$this->slug_name            = ! empty( $slug_name ) ? strtolower( $slug_name ) : $this->plugin_name;
-
-		$this->setup_library_version();
-		$this->setup_actions();
 	}
 
 	/**
@@ -261,12 +282,12 @@ class Technote {
 		} );
 
 		add_action( 'init', function () {
-			$this->initialize();
+			$this->main_init();
 		}, 1 );
 
 		add_action( 'activated_plugin', function ( $plugin ) {
 			$this->plugins_loaded();
-			$this->initialize();
+			$this->main_init();
 			if ( $this->define->plugin_base_name === $plugin ) {
 				$this->filter->do_action( 'app_activated', $this );
 			}
@@ -288,12 +309,9 @@ class Technote {
 		}
 		$this->plugins_loaded = true;
 
-		if ( ! function_exists( 'get_plugin_data' ) ) {
-			require_once ABSPATH . 'wp-admin/includes/plugin.php';
-		}
-		$this->plugin_data = get_plugin_data( $this->plugin_file, false, false );
-
-		spl_autoload_register( [ $this, 'load_class' ] );
+		spl_autoload_register( function ( $class ) {
+			return $this->get_main()->load_class( $class );
+		} );
 
 		$this->load_functions();
 	}
@@ -310,276 +328,6 @@ class Technote {
 	}
 
 	/**
-	 * @param bool $uninstall
-	 */
-	private function initialize( $uninstall = false ) {
-		if ( $this->initialized ) {
-			return;
-		}
-		$this->initialized = true;
-
-		$this->filter->do_action( 'app_initialize', $this );
-		$this->setup_property( $uninstall );
-		$this->setup_textdomain();
-		$this->setup_settings();
-		$this->filter->do_action( 'app_initialized', $this );
-	}
-
-	/**
-	 * @return bool
-	 */
-	public function has_initialized() {
-		return $this->initialized;
-	}
-
-	/**
-	 * @since 2.0.0
-	 * @return string
-	 */
-	public function get_library_directory() {
-		return self::$latest_library_directory;
-	}
-
-	/**
-	 * @since 2.0.0
-	 * @return string
-	 */
-	public function get_library_version() {
-		return self::$latest_library_version;
-	}
-
-	/**
-	 * @since
-	 * @return string
-	 */
-	public function get_plugin_version() {
-		return $this->plugin_data['Version'];
-	}
-
-	/**
-	 * @since 2.8.1 Added: setup social login, custom post filters
-	 *
-	 * @param bool $uninstall
-	 */
-	private function setup_property( $uninstall ) {
-		if ( $uninstall ) {
-			foreach ( $this->properties as $name => $class ) {
-				$this->$name;
-			}
-			$this->uninstall->get_class_list();
-		} else {
-			if ( $this->get_config( 'config', 'use_custom_post' ) ) {
-				$this->custom_post;
-			}
-			if ( $this->get_config( 'config', 'use_social_login' ) ) {
-				$this->social;
-			}
-		}
-	}
-
-	/**
-	 * @since 1.1.73
-	 * @since 2.1.0 Changed: load textdomain from plugin data
-	 * @return mixed
-	 */
-	public function get_text_domain() {
-		return $this->define->plugin_textdomain;
-	}
-
-	/**
-	 * setup textdomain
-	 */
-	private function setup_textdomain() {
-		if ( ! static::$lib_language_loaded ) {
-			static::$lib_language_loaded = true;
-			load_plugin_textdomain( $this->define->lib_textdomain, false, $this->define->lib_languages_rel_path );
-		}
-
-		$text_domain = $this->get_text_domain();
-		if ( ! empty( $text_domain ) ) {
-			load_plugin_textdomain( $text_domain, false, $this->define->plugin_languages_rel_path );
-		}
-	}
-
-	/**
-	 * setup settings
-	 * @since 2.1.0 Changed: set default value of check_update when the plugin is registered as official
-	 */
-	private function setup_settings() {
-		if ( defined( 'TECHNOTE_MOCK_REST_REQUEST' ) && TECHNOTE_MOCK_REST_REQUEST ) {
-			$this->setting->remove_setting( 'use_admin_ajax' );
-		}
-		if ( $this->loader->api->is_empty() ) {
-			$this->setting->remove_setting( 'use_admin_ajax' );
-			$this->setting->remove_setting( 'get_nonce_check_referer' );
-			$this->setting->remove_setting( 'check_referer_host' );
-		}
-		if ( ! empty( $this->plugin_data['PluginURI'] ) && $this->utility->starts_with( $this->plugin_data['PluginURI'], 'https://wordpress.org' ) ) {
-			$this->setting->edit_setting( 'check_update', 'default', false );
-		}
-		if ( ! $this->log->is_valid() ) {
-			$this->setting->remove_setting( 'save___log_term' );
-			$this->setting->remove_setting( 'delete___log_interval' );
-		}
-		if ( $this->get_config( 'config', 'prevent_use_log' ) ) {
-			$this->setting->remove_setting( 'is_valid_log' );
-			$this->setting->remove_setting( 'capture_shutdown_error' );
-		}
-	}
-
-	/**
-	 * @param $value
-	 *
-	 * @return string
-	 */
-	public function translate( $value ) {
-		$text_domain = $this->get_text_domain();
-		if ( ! empty( $text_domain ) ) {
-			$translated = __( $value, $text_domain );
-			if ( $value !== $translated ) {
-				return $translated;
-			}
-		}
-
-		return __( $value, $this->define->lib_textdomain );
-	}
-
-	/**
-	 * @param string $name
-	 * @param string $key
-	 * @param mixed $default
-	 *
-	 * @return mixed
-	 */
-	public function get_config( $name, $key, $default = null ) {
-		return $this->config->get( $name, $key, $default );
-	}
-
-	/**
-	 * @param string $key
-	 * @param mixed $default
-	 *
-	 * @return mixed
-	 */
-	public function get_option( $key, $default = '' ) {
-		return $this->option->get( $key, $default );
-	}
-
-	/**
-	 * @param string $key
-	 * @param mixed $default
-	 *
-	 * @return mixed
-	 */
-	public function get_session( $key, $default = null ) {
-		return $this->session->get( $key, $default );
-	}
-
-	/**
-	 * @param string $key
-	 * @param mixed $value
-	 * @param int|null $duration
-	 */
-	public function set_session( $key, $value, $duration = null ) {
-		$this->session->set( $key, $value, $duration );
-	}
-
-	/**
-	 * @param null|string|false $capability
-	 *
-	 * @return bool
-	 */
-	public function user_can( $capability = null ) {
-		return $this->user->user_can( $capability );
-	}
-
-	/**
-	 * @param string $message
-	 * @param mixed $context
-	 * @param string $level
-	 */
-	public function log( $message, $context = null, $level = '' ) {
-		if ( $message instanceof \Exception ) {
-			$this->log->log( $message->getMessage(), isset( $context ) ? $context : $message->getTraceAsString(), empty( $level ) ? 'error' : $level );
-		} elseif ( $message instanceof \WP_Error ) {
-			$this->log->log( $message->get_error_message(), isset( $context ) ? $context : $message->get_error_data(), empty( $level ) ? 'error' : $level );
-		} else {
-			$this->log->log( $message, $context, $level );
-		}
-	}
-
-	/**
-	 * @param string $message
-	 * @param string $group
-	 * @param bool $error
-	 * @param bool $escape
-	 */
-	public function add_message( $message, $group = '', $error = false, $escape = true ) {
-		if ( ! isset( $this->loader->admin ) ) {
-			add_action( 'admin_notices', function () use ( $message, $group, $error, $escape ) {
-				$this->loader->admin->add_message( $message, $group, $error, $escape );
-			}, 9 );
-		} else {
-			$this->loader->admin->add_message( $message, $group, $error, $escape );
-		}
-	}
-
-	/**
-	 * @param $class
-	 *
-	 * @return bool
-	 */
-	public function load_class( $class ) {
-		$class = ltrim( $class, '\\' );
-		$dir   = null;
-		if ( empty( $this->define ) ) {
-			$namespace = ucfirst( TECHNOTE_PLUGIN );
-			$class     = preg_replace( "#\A{$namespace}#", '', $class );
-			$dir       = self::$latest_library_directory . DS . 'src';
-		} elseif ( preg_match( "#\A{$this->define->plugin_namespace}#", $class ) ) {
-			$class = preg_replace( "#\A{$this->define->plugin_namespace}#", '', $class );
-			$dir   = $this->define->plugin_src_dir;
-		} elseif ( preg_match( "#\A{$this->define->lib_namespace}#", $class ) ) {
-			$class = preg_replace( "#\A{$this->define->lib_namespace}#", '', $class );
-			$dir   = $this->define->lib_src_dir;
-		}
-
-		if ( isset( $dir ) ) {
-			$class = ltrim( $class, '\\' );
-			$class = strtolower( $class );
-			$path  = $dir . DS . str_replace( '\\', DS, $class ) . '.php';
-			if ( is_readable( $path ) ) {
-				/** @noinspection PhpIncludeInspection */
-				require_once $path;
-
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/**
-	 * @param $file
-	 *
-	 * @return string
-	 */
-	public function get_page_slug( $file ) {
-		return basename( $file, '.php' );
-	}
-
-	/**
-	 * @param $name
-	 * @param $arguments
-	 */
-	public static function __callStatic( $name, $arguments ) {
-		if ( preg_match( '#register_uninstall_(.+)\z#', $name, $matches ) ) {
-			$plugin_base_name = $matches[1];
-			self::uninstall( $plugin_base_name );
-		}
-	}
-
-	/**
 	 * @since 2.0.2 Fixed: Uninstall behavior
 	 *
 	 * @param string $plugin_base_name
@@ -591,7 +339,7 @@ class Technote {
 		}
 
 		$app->plugins_loaded();
-		$app->initialize( true );
+		$app->main_init( true );
 		$app->uninstall->uninstall();
 	}
 
@@ -615,46 +363,19 @@ class Technote {
 	}
 
 	/**
-	 * @param string $key
-	 * @param string|null $target
-	 *
-	 * @return mixed
+	 * @since 2.0.0
+	 * @return string
 	 */
-	public function get_shared_object( $key, $target = null ) {
-		! isset( $target ) and $target = $this->plugin_name;
-
-		return isset( self::$shared_object[ $target ][ $key ] ) ? self::$shared_object[ $target ][ $key ] : null;
+	public function get_library_directory() {
+		return self::$latest_library_directory;
 	}
 
 	/**
-	 * @param string $key
-	 * @param mixed $object
-	 * @param string|null $target
+	 * @since 2.0.0
+	 * @return string
 	 */
-	public function set_shared_object( $key, $object, $target = null ) {
-		! isset( $target ) and $target = $this->plugin_name;
-		self::$shared_object[ $target ][ $key ] = $object;
-	}
-
-	/**
-	 * @param string $key
-	 * @param string|null $target
-	 *
-	 * @return bool
-	 */
-	public function isset_shared_object( $key, $target = null ) {
-		! isset( $target ) and $target = $this->plugin_name;
-
-		return isset( self::$shared_object[ $target ] ) && array_key_exists( $key, self::$shared_object[ $target ] );
-	}
-
-	/**
-	 * @param string $key
-	 * @param string|null $target
-	 */
-	public function delete_shared_object( $key, $target = null ) {
-		! isset( $target ) and $target = $this->plugin_name;
-		unset( self::$shared_object[ $target ][ $key ] );
+	public function get_library_version() {
+		return self::$latest_library_version;
 	}
 }
