+++ src/traits/singleton.php
@@ -133,7 +133,6 @@ trait Singleton {
 	 * @return string
 	 */
 	public function get_slug( $config_name, $suffix = '-' ) {
-
 		if ( ! isset( self::$slugs[ $this->app->plugin_name ][ $config_name ] ) ) {
 			$default = $this->app->slug_name . $suffix;
 			$slug    = $this->app->get_config( 'slug', $config_name, $default );
